Ensure node-only exports are only available in node environments
 * Separates NodeWallet into its own file
 * Ensures that NodeWallet and workspace are only exported on Node
